add optional

+++ src/main/java/ru/sbrf/java/payment/appendix/operation/ToPhonePayment.java
@@ -3,7 +3,9 @@ package ru.sbrf.java.payment.appendix.operation;
 import ru.sbrf.java.payment.client.Currency;
 import ru.sbrf.java.payment.client.User;
 
+import java.math.BigDecimal;
+
 public interface ToPhonePayment {
 
-    void payToPhone(long targetNumber, long sum, Currency currency);
+    void payToPhone(long targetNumber, BigDecimal sum, Currency currency);
 }
